Call allele specific copy number in single cell datasets

Call allele specific copy number in single cell datasets

callAlleleSpecificCN(CNbins, haplotypes, eps = 1e-12, loherror = 0.02,
  maxCN = NULL, selftransitionprob = 0.999, progressbar = TRUE,
  ncores = 1, likelihood = "binomial", minbins = 100,
  minbinschr = 10)

Arguments

CNbins
single cell copy number dataframe with the following columns: cell_id, chr, start, end, state, copy

haplotypes
single cell haplotypes dataframe with the following columns: cell_id, chr, start, end, hap_label, allele1, allele0, totalcounts

eps
default 1e-12

loherror
LOH error rate for initial assignment, this is inferred directly from the data in the second pass, default = 0.02

maxCN
maximum copy number to infer allele specific states, default=NULL which will use the maximum state from CNbins

selftransitionprob
probability to stay in the same state in the HMM, default = 0.999, set to 0.0 for an IID model

progressbar
Boolean to display progressbar or not, default = TRUE, will only show if ncores == 1

ncores
Number of cores to use, default = 1

likelihood
Likelihood model for HMM, default is binomial, other option is betabinomial or use auto and the algorithm will choose the likelihood that best fits the data.

minbins
Minimum number of bins containing both haplotype counts and copy number data for a cell to be included

minbinschr
Minimum number of bins containing both haplotype counts and copy number data per chromosome for a cell to be included

Value

allele specific copy number object which includes dataframe similar to input with additional columns which include

Maj (Major allele copy number)

Min (Minor allele copy number)

state_AS_phased (phased state of the form Maj|Min )

state_AS (mirrored state of the form Maj|Min)

LOH (is bin LOH or not)

state_phase (state describing which is the dominant allele and whether it is LOH or not)

state_BAF (binned discretized BAF value calculated as Min / (Maj + Min))

Details

In the allele specific copy number inference Maj is always > Min and state_AS_phased == state_AS

Examples

sim_data <- simulate_data_cohort(clone_num = c(20, 20),
       clonal_events = list(list("1" = c(2,0), "5" = c(3,1)),
                       list("2" = c(6,3), "3" = c(1,0))),
       loherror = 0.02,
       coverage = 30)

#> Joining, by = c("chr", "start", "end", "hap_label")

results <- callAlleleSpecificCN(sim_data$CNbins, sim_data$haplotypes)

#> Finding overlapping cell IDs between CN data and haplotype data...

#> Total number of cells in both CN and haplotypes: 40

#> Number of cells in CN data: 40

#> Removing cells from CN data...

#> Number of cells in haplotype data: 40

#> Removing cells from haplotype data...

#> Reformatting haplotypes

#> Pivot data frame...

#> Phase haplotypes...

#> Phasing based on distribution across all cells

#> Join phased haplotypes...

#> Reorder haplotypes based on phase...

#> Joining bins and haplotypes...

#> Calculate BAF per bin...

#> Total number of cells after removing cells with < 100 bins: 40

#> Initial assignment...

#> Rerun using mirrored BAF...

#> Average distance from median to expected BAF = 0.0077