Call haplotype specific copy number in single cell datasets

Call haplotype specific copy number in single cell datasets

callHaplotypeSpecificCN(
  CNbins,
  haplotypes,
  eps = 1e-12,
  loherror = 0.02,
  maxCN = NULL,
  selftransitionprob = 0.95,
  progressbar = TRUE,
  ncores = 1,
  phasebyarm = FALSE,
  minfrac = 0.8,
  likelihood = "binomial",
  minbins = 100,
  minbinschr = 10,
  phased_haplotypes = NULL,
  clustering_method = "copy",
  maxloherror = 0.035,
  overwritemincells = NULL,
  cluster_per_chr = FALSE,
  viterbiver = "cpp"
)

Arguments

CNbins
single cell copy number dataframe with the following columns: cell_id, chr, start, end, state, copy

haplotypes
single cell haplotypes dataframe with the following columns: cell_id, chr, start, end, hap_label, allele1, allele0, totalcounts

eps
default 1e-12

loherror
LOH error rate for initial assignment, this is inferred directly from the data in the second pass, default = 0.02

maxCN
maximum copy number to infer allele specific states, default=NULL which will use the maximum state from CNbins

selftransitionprob
probability to stay in the same state in the HMM, default = 0.95, set to 0.0 for an IID model

progressbar
Boolean to display progressbar or not, default = TRUE, will only show if ncores == 1

ncores
Number of cores to use, default = 1

minfrac
Minimum proportion of haplotypes to retain when clustering + phasing, default = 0.8

likelihood
Likelihood model for HMM, default is binomial, other option is betabinomial or use auto and the algorithm will choose the likelihood that best fits the data.

minbins
Minimum number of bins containing both haplotype counts and copy number data for a cell to be included

minbinschr
Minimum number of bins containing both haplotype counts and copy number data per chromosome for a cell to be included

phased_haplotypes
Use this if you want to manually define the haplotypes phasing if for example the default heuristics used by schnapps does not return a good fit.

clustering_method
Method to use to cluster cells for haplotype phasing, default is copy, other option is breakpoints

maxloherror
Maximum value for LOH error rate

overwritemincells
default NULL

cluster_per_chr
Whether to cluster per chromosome to rephase alleles or not

Value

allele specific copy number object which includes dataframe similar to input with additional columns which include

Maj (Major allele copy number)

Min (Minor allele copy number)

state_AS_phased (phased state of the form Maj|Min )

state_AS (mirrored state of the form Maj|Min)

LOH (is bin LOH or not)

state_phase (state describing which is the dominant allele and whether it is LOH or not)

state_BAF (binned discretized BAF value calculated as Min / (Maj + Min))

Examples

sim_data <- simulate_data_cohort(
  clone_num = c(20, 20),
  clonal_events = list(
    list("1" = c(2, 0), "5" = c(3, 1)),
    list("2" = c(6, 3), "3" = c(1, 0))
  ),
  loherror = 0.02,
  coverage = 100
)

#> Joining, by = c("chr", "start", "end", "hap_label")

results <- callHaplotypeSpecificCN(sim_data$CNbins, sim_data$haplotypes)

#> Finding overlapping cell IDs between CN data and haplotype data...

#> Total number of cells in both CN and haplotypes: 40

#> Number of cells in CN data: 40

#> Number of cells in haplotype data: 40

#> Reformatting haplotypes

#> Phase haplotypes...

#> Phasing based on distribution across all cells

#> Join phased haplotypes...

#> Reorder haplotypes based on phase...

#> Joining bins and haplotypes...

#> Calculate BAF per bin...

#> Total number of cells after removing cells with < 100 bins: 40

#> Using 2 cells for clustering...

#> Creating CN matrix...

#> Calculating UMAP dimensionality reduction...

#> Clustering cells using hdbscan...

#> Identified 2 clusters

#> Distribution of clusters:

#>   Cluster A: 20

#>   Cluster B: 20

#> Joining, by = "cell_id"

#> Phasing by chromosome (not arm level)..

#> Finding overlapping cell IDs between CN data and haplotype data...

#> Total number of cells in both CN and haplotypes: 40

#> Number of cells in CN data: 40

#> Number of cells in haplotype data: 40

#> Reformatting haplotypes

#> Phase haplotypes...

#> Join phased haplotypes...

#> Reorder haplotypes based on phase...

#> Joining bins and haplotypes...

#> Calculate BAF per bin...

#> Total number of cells after removing cells with < 100 bins: 40

#> Average distance from median to expected BAF = 0.0037

#> Average distance from median to expected BAF = 0.0037