Heatmap plot

Plot a heatmap where rows are cells, columns are genome coordinates and colours map to (allele-specific) copy-number states

plotHeatmap(
  cn,
  tree = NULL,
  clusters = NULL,
  normalize_ploidy = FALSE,
  normalize_tree = FALSE,
  branch_length = 1,
  spacer_cols = 20,
  plottree = TRUE,
  plotcol = "state",
  reorderclusters = FALSE,
  pctcells = 0.05,
  library_mapping = NULL,
  clone_pal = NULL,
  sample_label_idx = 1,
  fillna = TRUE,
  frequencycutoff = 2,
  maxf = NULL,
  plotfrequency = FALSE,
  show_legend = TRUE,
  show_library_label = TRUE,
  show_clone_label = TRUE,
  widenarm = FALSE,
  umapmetric = "euclidean",
  chrlabels = TRUE,
  raster_quality = 10,
  SV = NULL,
  seed = NULL,
  nticks = 4,
  fillgenome = FALSE,
  annotation_height = NULL,
  annofontsize = 14,
  na_col = "white",
  linkheight = 5,
  newlegendname = NULL,
  ...
)

Arguments

cn
Either a hscn object or a single cell allele specific copy number dataframe with the following columns: `cell_id`, `chr`, `start`, `end`, `state`, `copy`

tree
Tree in newick format to plot alongside the heatmap, default = NULL

normalize_ploidy
Normalize ploidy of all cells to 2

normalize_tree
default = FALSE

branch_length
scales branch lengths to this size, default = 2

spacer_cols
number of empty columns between chromosomes, default = 20

plottree
Binary value of whether to plot tree or not, default = TRUE

plotcol
Which column to colour the heatmap by, should be one of "state", "state_BAF", "state_phase", "state_AS", "state_min", "copy", "BAF", "Min", "Maj"

pctcells
Minimum size of cluster in terms of perecentage of cells in umap clustering

library_mapping
Named vector mapping library names to labels for legend

clone_pal
pallette to colour clusters by

sample_label_idx
default = 1

fillna
Smooth over NA values, default = TRUE

frequencycutoff
default = 2

maxf
Max frequency when plotting the frequency track, default = NULL infers this from the data

plotfrequency
Plot the frequency track of gains and losses across the genome

show_legend
plot legend or not, boolean

show_library_label
show library label or not, boolean

show_clone_label
show clone label or not, boolean

umapmetric
metric to use in umap dimensionality reduction if no clusters are specified

chrlabels
include chromosome labels or not, boolean

SV
sv data frame

seed
seed for UMAP

nticks
number of ticks in x-axis label when plotting a single chromosome

fillgenome
fill in any missing bins and add NA to centromeric regions

na_col
colour of NA values

linkheight
height of x-axis ticks

newlegendname
overwrite default legend name

If clusters are set to NULL then the function will compute clusters using UMAP and HDBSCAN.

#'

cluster
data.frame assigning cells to clusters, needs the following columns `cell_id`, `clone_id` default = NULL

reordercluster
Reorder the cells according to cluster if no tree is specified