shiny server relevant functions

shiny server relevant functions

schnapps(
  localContributionDir = "~/Rstudio/shHubgit/Dummy/",
  defaultValueSingleGene = "CD3g",
  defaultValueMultiGenes =
    "cd3g, cd4, cd8b, ms4a1, TCF4, LILRA2, LYZ, cd79a, bcl11b, IL32, hbb, nkg7,MNDA",
  defaultValueRegExGene = "",
  DEBUG = FALSE,
  DEBUGSAVE = FALSE,
  historyPath = NULL,
  defaultValues = list(),
  port = NULL,
  launch.browser = getOption("shiny.launch.browser", interactive())
)

Arguments

localContributionDir

path to the directory(ies) that contain additional functionality

defaultValueSingleGene

single gene name to used as a default value.

defaultValueMultiGenes

comma separated list of gene names to be used as a default value.

defaultValueRegExGene

regular Expression used for gene selection

DEBUG

TRUE/FALSE whether to show debugging information on the console

DEBUGSAVE

TRUE/FALSE where or not save internal data (very time consuming)

historyPath

location (directory) where history directories and data will be stored.

defaultValues

list of default values to use for inputs

launch.browser

a function to call with the application's URL.

Details

Shiny app for the analysis of single cell data

Examples

# create example data
data("scEx", package = "SCHNAPPs")
save(file = "scEx.Rdata", list = "scEx")
#> Error in save(file = "scEx.Rdata", list = "scEx"): object 'scEx' not found
# use "scEx.Rdata" with load data functionality within the shiny app