shiny server relevant functions

shiny server relevant functions

schnappsLite(
  data = "RdataFile",
  localContributionDir = "~/Rstudio/shHubgit/Dummy/",
  defaultValueSingleGene = "CD52",
  defaultValueMultiGenes = "CD52, S100A4, S100A9, S100A8",
  defaultValueRegExGene = "",
  DEBUG = FALSE,
  DEBUGSAVE = FALSE,
  historyPath = NULL,
  maxCells = 3000,
  defaultValues = list(),
  AllowClustering = FALSE
)

Arguments

data

RdataFile data file with scEx, scEx_log, projections

localContributionDir

path to the directory(ies) that contain additional functionality

defaultValueSingleGene

single gene name to used as a default value.

defaultValueMultiGenes

comma separated list of gene names to be used as a default value.

defaultValueRegExGene

regular Expression used for gene selection

DEBUG

TRUE/FALSE whether to show debugging information on the console

DEBUGSAVE

TRUE/FALSE where or not save internal data (very time consuming)

historyPath

location (directory) where history directories and data will be stored.

maxCells

maximal number of cells to use. Used to limit memory usage for server

defaultValues

list of default values to use for inputs

AllowClustering

whether to include functionality to cluster cells.

Details

Shiny app for the analysis of single cell data

Examples

packPath = "schnappsPackage"
# TODO fix:
# schnappsLite(data = paste0(packPath, "/data/scExLite.RData"))